[REFACTOR] Infrastructure repository 하위 패키지 이동
