+++ src/main/java/kr/hhplus/be/ecommerce/infrastructure/payment/repository/PaymentJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.ecommerce.infrastructure.payment;
+package kr.hhplus.be.ecommerce.infrastructure.payment.repository;
 
 import kr.hhplus.be.ecommerce.domain.payment.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
