+++ src/main/java/kr/hhplus/be/ecommerce/infrastructure/payment/repository/PaymentRepositoryImpl.java
@@ -1,11 +1,9 @@
-package kr.hhplus.be.ecommerce.infrastructure.payment;
+package kr.hhplus.be.ecommerce.infrastructure.payment.repository;
 
 import kr.hhplus.be.ecommerce.domain.payment.Payment;
 import kr.hhplus.be.ecommerce.domain.payment.PaymentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-
-import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
@@ -19,7 +17,8 @@ public class PaymentRepositoryImpl implements PaymentRepository {
     }
 
     @Override
-    public Optional<Payment> findById(Long id) {
-        return paymentJpaRepository.findById(id);
+    public Payment findById(Long id) {
+        return paymentJpaRepository.findById(id)
+            .orElseThrow(() -> new IllegalArgumentException("결제 정보를 찾을 수 없습니다."));
     }
 }
